+++ src/Pages/Mypage/Profile.jsx
@@ -11,6 +11,7 @@ const Profile = () => {
   const [image, setImage] = useState(null);
   const [editingNickname, setEditingNickname] = useState(false);
   const [newNickname, setNewNickname] = useState("");
+  const [nicknameError, setNicknameError] = useState("");
 
   // input 요소에 대한 참조
   const fileInputRef = useRef(null);
@@ -34,14 +35,20 @@ const Profile = () => {
 
   // 닉네임 변경 핸들러
   const handleNicknameChange = (e) => {
-    setNewNickname(e.target.value);
-    setNickname(e.target.value);
+    const newNickname = e.target.value;
+    setNewNickname(newNickname);
+    setNicknameError(validateNickname(newNickname));
   };
 
   // 닉네임 저장 핸들러
   const handleSaveNickname = () => {
-    setNickname(newNickname);
-    setEditingNickname(false);
+    const error = validateNickname(newNickname);
+    if (error) {
+      setNicknameError(error);
+    } else {
+      setNickname(newNickname);
+      setEditingNickname(false);
+    }
   };
 
   // 성별 변경 핸들러
@@ -56,7 +63,21 @@ const Profile = () => {
 
   const handleButtonClick = () => {
     alert("저장되었습니다!");
-};
+  };
+
+  // 닉네임 유효성 검사
+  const eng = /[a-zA-Z]/;
+  const kor = /[ㄱ-ㅎ|ㅏ-ㅣ|가-힣]/;
+  const num = /[0-9]/;
+  const spc = /[~!@#$%^&*()_+|<>?:{}]/;
+
+  const validateNickname = ({ target: { value: input } }) => {
+    if (input.length >= 0 && input.length <= 12 && (eng.test(input) || kor.test(input) || num.test(input) || !spc.test(input))) {
+      setNickname(input);
+    } else {
+      alert("닉네임은 영문, 한글, 숫자를 포함한 12글자 이하여야 하며 특수기호를 포함하지 않아야 합니다.");
+    }
+  };
 
   return (
     <div className="ProfileWrap">
@@ -85,7 +106,6 @@ const Profile = () => {
                 />
               )}
               <div>
-                {/* 숨겨진 파일 입력 */}
                 <input
                   type="file"
                   accept="image/*"
@@ -98,14 +118,26 @@ const Profile = () => {
           </div>
 
           <div className="Profileinputnic">
-            <input
-              type="text"
-              value={nickname}
-              onChange={handleNicknameChange}
-              onBlur={handleNicknameChange}
-              className="Profilenickname"
-            />
-        </div>
+            {!editingNickname ? (
+              <div className="Profilenickname" onClick={handleEditNickname}>
+                {nickname}
+              </div>
+            ) : (
+              <div>
+                <input
+                  type="text"
+                  value={newNickname}
+                  onChange={handleNicknameChange}
+                  onBlur={handleNicknameChange}
+                  className="Profilenickname"
+                />
+                {nicknameError && (
+                  <div className="Error">{nicknameError}</div>
+                )}
+                <button onClick={handleSaveNickname}>저장</button>
+              </div>
+            )}
+          </div>
 
           <div className="Profileinput">
             <label>성별</label>
@@ -124,12 +156,13 @@ const Profile = () => {
             <input
               type="number"
               value={age}
-              값호출
               onChange={handleAgeChange}
               className="Profileage"
             />
           </div>
-          <button className="Profilesubmit" onClick={handleButtonClick}>저장</button>
+          <button className="Profilesubmit" onClick={handleButtonClick}>
+            저장
+          </button>
         </div>
       </div>
     </div>
